+++ mordant/src/appleMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.apple.kt
@@ -1,6 +1,6 @@
 package com.github.ajalt.mordant.internal
 
-import com.github.ajalt.mordant.internal.syscalls.SyscallHandler
-import com.github.ajalt.mordant.internal.syscalls.SyscallHandlerNativeApple
+import com.github.ajalt.mordant.terminal.TerminalInterface
+import com.github.ajalt.mordant.internal.syscalls.SyscallHandlerNativePosix
 
-internal actual fun getSyscallHandler(): SyscallHandler = SyscallHandlerNativeApple
+internal actual fun getStandardTerminalInterface(): TerminalInterface = SyscallHandlerNativeApple