+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/widgets/progress/CachedProgressBarDefinition.kt
@@ -4,7 +4,6 @@ import com.github.ajalt.mordant.internal.MppAtomicRef
 import com.github.ajalt.mordant.internal.update
 import com.github.ajalt.mordant.rendering.Widget
 import kotlin.time.ComparableTimeMark
-import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
 import kotlin.time.TimeSource